test: don't bother disconnecting workers
cluster.disconnect() doesn't complete on node v0.12+. I'm not sure why,
but I suspect that HTTP connections to the workers are being kept alive.
The workers won't exit until the servers have closed, and the servers in
the workers won't close until all incoming connections are done. We
could figure out why this is different, but disconnecting the workers is
unnecessary - they exit on master exit, so just delete the
cluster.disconnect().